backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pandas
numpy
scikit-learn
kaggle
pydantic
python-multipart
requests
